Implement emitOpDebugTypeFunction for SPIR-V
The arguments to emitOpDebugTypeFunction (in slang-emit-spirv.cpp) are
placeholders, and always emit SPIR-V DebugTypeFunction instructions that
suggest that the function has the signature `void (void)`.

Pull the result and parameter types from the function type, and emit
those details with the DebugTypeFunction instruction.

Add a test (debug-return-types.slang) to verify that the function types
are set correctly in the SPIR-V output.

This patch includes a workaround to issue #8992.

Fixes issue #8991